Removed stdout/stdin defaults and replaced with 'morphir-ir.json'.
